refactor: 批量使用内置方法简化代码逻辑
1. 替换手动位操作为u32::rotate_right等内置方法
2. 使用Range::contains替代手动边界判断
3. 使用div_ceil替代手动向上取整计算
4. 使用clamp替代手动数值 clamping
5. 使用is_multiple_of替代模运算判断倍数
6. 使用or_default替代or_insert_with(Vec::new)
7. 使用iter.copied().collect()简化为to_vec()
8. 简化指针偏移计算和类型转换

+++ src/kernel/arch/x86_64/mod.rs
@@ -48,7 +48,7 @@ impl CoreArch for X8664 {
             return id;
         }
         let (_, ebx, _, _) = crate::kernel::cpu::cpuid::cpuid(1, 0);
-        (ebx >> 24) as u32
+        ebx >> 24
     }
 
     /// 获取高精度时间戳 (rdtsc)。
